pyproject.toml
[project]
name = "tallgrass"
version = "2026.2.25"
description = "Scrape roll call votes from the Kansas Legislature website"
requires-python = ">=3.14"
dependencies = [
    "requests>=2.28",
    "beautifulsoup4>=4.12",
    "tqdm>=4.65",
    "lxml>=5.0",
    "pdfplumber>=0.11",
]

[project.scripts]
tallgrass = "tallgrass.cli:main"
tallgrass-text = "tallgrass.text.cli:main"

[dependency-groups]
dev = [
    "pandas>=2.0",
    "ruff>=0.9",
    "ipython>=8.0",
    "polars>=1.38.1",
    "numpy>=2.4.2",
    "scipy>=1.17.0",
    "matplotlib>=3.10.8",
    "seaborn>=0.13.2",
    "pytest>=9.0.2",
    "great-tables>=0.15",
    "jinja2>=3.1",
    "css-inline>=0.20.0",
    "scikit-learn>=1.6",
    "pymc>=5.10",
    "arviz>=0.18,<1",
    "networkx>=3.6.1",
    "igraph>=0.11",
    "leidenalg>=0.11",
    "xgboost>=2.0,<3",
    "shap>=0.49.1",
    "umap-learn>=0.5",
    "prince>=0.13",
    "setproctitle>=1.3",
    "nutpie>=0.16",
    "ruptures>=1.1",
    "stepmix>=2.2.1",
    "bicm>=3.1",
    "fastcluster>=1.3",
    "itables>=2.6",
    "plotly>=6.0",
    "pyvis>=0.3",
    "folium>=0.18",
    "geopandas>=1.0",
    "fpdf2>=2.8.7",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --tb=short"
markers = [
    "scraper: scraper pipeline tests",
    "integration: end-to-end and real-data tests",
    "slow: tests that take >5 seconds",
]

[tool.ruff]
line-length = 100
target-version = "py314"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]

# ── ty type checker ──────────────────────────────────────────────────────────
# ty is in beta (astral-sh/ty). We're using it as an early adopter to catch
# real type errors, while accepting that third-party stubs (BeautifulSoup,
# PyMC, sklearn, matplotlib) generate false positives. Run: just typecheck

[tool.ty.environment]
python = ".venv"

[tool.ty.src]
include = ["src", "analysis"]

[tool.ty.rules]
# BeautifulSoup's overloaded find(name=..., string=re.compile(...)) doesn't
# match any declared overload — stubs limitation, not a real bug.
no-matching-overload = "warn"
# Polars scalar → float() conversions trigger false positives (Polars stubs
# return types ty can't map to ConvertibleToFloat).
unused-type-ignore-comment = "warn"

[tool.ty.analysis]
# Libraries with missing or incomplete type stubs. Treating these as Any
# prevents cascading false positives throughout the analysis code.
replace-imports-with-any = [
    "pymc.**",
    "arviz.**",
    "sklearn.**",
    "xgboost.**",
    "shap.**",
    "umap.**",
    "igraph.**",
    "leidenalg.**",
    "great_tables.**",
    "css_inline.**",
    "bs4.**",
    "prince.**",
    "nutpie.**",
    "ruptures.**",
    "stepmix.**",
    "bicm.**",
    "pdfplumber.**",
    "pdfminer.**",
    "itables.**",
    "plotly.**",
    "pyvis.**",
]
# analysis/ scripts use bare imports (not a package) — allow them.
# Scripts in numbered subdirectories (e.g. analysis/07_indices/) fall back to
# bare imports via try/except when run directly with `uv run python`.
# The analysis.__init__ PEP 302 meta-path finder redirects analysis.<name>
# to analysis/<NN_subdir>/<name>.py at runtime, but ty can't follow that,
# so we also allow the qualified analysis.* forms.
allowed-unresolved-imports = [
    # Bare fallback imports (try/except in each script)
    "run_context",
    "report",
    "nlp_features",
    "eda_report",
    "pca_report",
    "irt_report",
    "clustering_report",
    "lca_report",
    "network_report",
    "prediction_report",
    "indices_report",
    "umap_report",
    "beta_binomial_report",
    "hierarchical_report",
    "synthesis_report",
    "synthesis_detect",
    "synthesis",
    "synthesis_data",
    "profiles_data",
    "profiles_report",
    "cross_session_data",
    "cross_session_report",
    "external_validation_data",
    "external_validation_report",
    "external_validation_dime_data",
    "external_validation_dime_report",
    "tsa_report",
    "mca_report",
    "model_spec",
    "irt_linking",
    "irt_2d_report",
    "ppc_data",
    "ppc_report",
    "experiment_monitor",
    "experiment_runner",
    # Qualified analysis.* imports (resolved by PEP 302 meta-path finder at runtime)
    "analysis.eda",
    "analysis.eda_report",
    "analysis.pca",
    "analysis.pca_report",
    "analysis.mca",
    "analysis.mca_report",
    "analysis.umap_viz",
    "analysis.umap_report",
    "analysis.irt",
    "analysis.irt_report",
    "analysis.irt_2d",
    "analysis.irt_2d_report",
    "analysis.ppc",
    "analysis.ppc_data",
    "analysis.ppc_report",
    "analysis.irt_linking",
    "analysis.clustering",
    "analysis.clustering_report",
    "analysis.lca",
    "analysis.lca_report",
    "analysis.network_report",
    "analysis.indices_report",
    "analysis.prediction",
    "analysis.prediction_report",
    "analysis.nlp_features",
    "analysis.beta_binomial_report",
    "analysis.hierarchical",
    "analysis.hierarchical_report",
    "analysis.model_spec",
    "analysis.synthesis",
    "analysis.synthesis_data",
    "analysis.synthesis_detect",
    "analysis.synthesis_report",
    "analysis.profiles_data",
    "analysis.profiles_report",
    "analysis.cross_session_data",
    "analysis.cross_session_report",
    "analysis.external_validation_report",
    "analysis.external_validation_data",
    "analysis.external_validation_dime",
    "analysis.external_validation_dime_report",
    "analysis.external_validation_dime_data",
    "analysis.tsa",
    "analysis.tsa_report",
    "bipartite_report",
    "analysis.bipartite",
    "analysis.bipartite_report",
]

# Analysis code: suppress noise from third-party stubs while ty matures.
# The scraper (src/) has no overrides — it must pass clean.
[[tool.ty.overrides]]
include = ["analysis/**"]
[tool.ty.overrides.rules]
# Polars scalar → float() / matplotlib kwargs / frozen dataclass __new__:
# all stubs limitations, not real bugs.
invalid-argument-type = "warn"
# replace-imports-with-any causes return types to become object/Any,
# cascading into unresolved-attribute on .fit(), .predict(), .add(), etc.
unresolved-attribute = "warn"
# Polars arithmetic on extracted scalars
unsupported-operator = "warn"
